reset error page when user navigates
This lets the user navigate away from the error. For example, if an error occurs on one tab in the user area, they can click on other tabs and those will work. Previously, the error would remain even when the user attempted to navigate away from it, as long as the error's parent stayed the same.